Don't swallow exception if messageArrivedAction fails
Inform end users about problems if `messageArrivedAction` fails, this would commonly occur if there is an exception thrown inside of a `messageArrived` callback leading to very hard to track silent errors. I'm not 100% on the functionality of `mqttService.traceError`, but if it ends up in Logcat this PR works as intended.