Use AndroidHighResolutionTimer for elapsed time
This timer uses SystemClock.elapsedRealtimeNanos to calculate the high
resolution timing value, which advances time when the device is asleep.

Fixes a bug introduced by the paho.mqtt.java library when it was changed
to track elapsed time using System.nanoTime. However, on android,
nanoTime does not advance when the device is in a deep sleep. As a
result, pings may not be sent frequently enough, resulting in client
disconnects.

See:
https://github.com/eclipse/paho.mqtt.java/issues/278
https://github.com/eclipse/paho.mqtt.java/issues/774

Signed-off-by: Dustin Thomson <dthomson@51systems.com>